[refactor] 検索画面で使うアダプタを別ファイルに移動
